+++ internal/serve/serve.go
@@ -21,7 +21,6 @@ import (
 	"github.com/stellar/stellar-disbursement-platform-backend/internal/circle"
 	"github.com/stellar/stellar-disbursement-platform-backend/internal/crashtracker"
 	"github.com/stellar/stellar-disbursement-platform-backend/internal/data"
-	"github.com/stellar/stellar-disbursement-platform-backend/internal/events"
 	"github.com/stellar/stellar-disbursement-platform-backend/internal/message"
 	"github.com/stellar/stellar-disbursement-platform-backend/internal/monitor"
 	"github.com/stellar/stellar-disbursement-platform-backend/internal/serve/httpclient"
@@ -87,11 +86,9 @@ type ServeOptions struct {
 	DisableMFA                      bool
 	DisableReCAPTCHA                bool
 	PasswordValidator               *authUtils.PasswordValidator
-	EnableScheduler                 bool // Deprecated: Use EventBrokerType=SCHEDULER instead.
 	tenantManager                   tenant.ManagerInterface
 	DistributionAccountService      services.DistributionAccountServiceInterface
 	DistAccEncryptionPassphrase     string
-	EventProducer                   events.Producer
 	MaxInvitationResendAttempts     int
 	SingleTenantMode                bool
 	CircleService                   circle.ServiceInterface
@@ -301,7 +298,6 @@ func handleHTTP(o ServeOptions) *chi.Mux {
 				DisbursementManagementService: &services.DisbursementManagementService{
 					Models:                     o.Models,
 					AuthManager:                authManager,
-					EventProducer:              o.EventProducer,
 					CrashTrackerClient:         o.CrashTrackerClient,
 					DistributionAccountService: o.DistributionAccountService,
 				},
@@ -343,12 +339,10 @@ func handleHTTP(o ServeOptions) *chi.Mux {
 				Models:                      o.Models,
 				DBConnectionPool:            o.MtnDBConnectionPool,
 				AuthManager:                 o.authManager,
-				EventProducer:               o.EventProducer,
 				CrashTrackerClient:          o.CrashTrackerClient,
 				DistributionAccountResolver: o.SubmitterEngine.DistributionAccountResolver,
 				DirectPaymentService: services.NewDirectPaymentService(
 					o.Models,
-					o.EventProducer,
 					o.DistributionAccountService,
 					o.SubmitterEngine,
 				),
@@ -405,7 +399,6 @@ func handleHTTP(o ServeOptions) *chi.Mux {
 			receiverWalletHandler := httphandler.ReceiverWalletsHandler{
 				Models:             o.Models,
 				CrashTrackerClient: o.CrashTrackerClient,
-				EventProducer:      o.EventProducer,
 			}
 
 			r.With(middleware.RequirePermission(
@@ -625,7 +618,6 @@ func handleHTTP(o ServeOptions) *chi.Mux {
 			ServiceID:        ServiceID,
 			Version:          o.Version,
 			DBConnectionPool: o.AdminDBConnectionPool,
-			Producer:         o.EventProducer,
 		}.ServeHTTP)
 
 		// START SEP-24 endpoints
@@ -665,7 +657,6 @@ func handleHTTP(o ServeOptions) *chi.Mux {
 				ReCAPTCHAValidator:          reCAPTCHAValidator,
 				ReCAPTCHADisabled:           o.DisableReCAPTCHA,
 				NetworkPassphrase:           o.NetworkPassphrase,
-				EventProducer:               o.EventProducer,
 				CrashTrackerClient:          o.CrashTrackerClient,
 				DistributionAccountResolver: o.SubmitterEngine.DistributionAccountResolver,
 			}.VerifyReceiverRegistration)